requirements.txt
faiss-cpu==1.9.0.post1
sentence-transformers==2.7.0
numpy==2.1.2
tqdm==4.66.4
python-docx==1.1.2
langchain-community==0.3.14
langchain-openai==0.2.10
langchain-huggingface==0.1.0
pywin32==311
requests==2.32.3